Add files via upload
Another go at ArxivReader.py This version has two abilities one for searching and returning results similar to how searxng.py returns results in a hope that it will make it ubiquitous and a download function that also reads files into memory.  This download function can download multiple articles based on the article id's being given to it with a comma in between.